requirements.txt
flask
flask-socketio
psutil
werkzeug
PyJWT
